+++ client/src/pages/Notes/canvasEngine/index.ts
@@ -1,4 +1,9 @@
 export * from './engineModel';
 export * from './geometry';
+export * from './measurementService';
 export * from './NoteCanvasRuntimeProvider';
+export * from './pageFrameService';
+export * from './placementService';
+export * from './runtimeLayout';
+export * from './viewportService';
 export type * from './types';